refactor(api-go): flat-events swap — sync cluster onto wire bindings
ChannelSync{Started,Progress,Completed} payloads become 'type X = wire.X'
aliases; hand-rolled envelopes deleted in the same commit; publishers construct
types.NewIntegrationEvent(wire.<X>EventName, …). Disclosed type adaptation:
percent cast uint32→int32 at the derivation site (binding types it int32,
0-100); HistorySyncType is already the exact-match wire alias — cast-free.

Wire-identity: 22 goldens byte-identical (goldens untouched); gateway
openapi.json regen byte-identical; go build/vet/test 12 pkgs ok.

+++ packages/api/go/internal/channel/handlers/channel_sync_handler.go
@@ -74,7 +74,7 @@ func (h *SyncProgressHandler) Handle(ctx context.Context, event types.DomainEven
 		ChannelID:       e.Payload.ChannelID,
 		OwnerID:         e.OwnerID,
 		HistorySyncType: e.Payload.HistorySyncType,
-		Percent:         e.Payload.Percent,
+		Percent:         int32(e.Payload.Percent),
 	})
 	if err := h.domainEventRepo.Save(ctx, domain); err != nil {
 		slog.Warn("failed to save sync_progress event",
